pyproject.toml
[project]
name = "polymarket-lp-tool"
version = "0.1.0"
description = "Polymarket passive liquidity rewards bot — monitor & auto-adjust limit orders"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client==0.34.6",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "python-socks[asyncio]>=2.8.1",
]

[project.scripts]
polymarket-lp = "passive_liquidity.main_loop:main"

[tool.uv]
dev-dependencies = []
